endret searchbar til klasse og flyttet states inn dit for å unngå unødvendig rerendring som treget til ting
