+++ src/Forvaltningsportalen/SearchBar/SearchBar.js
@@ -1,80 +1,205 @@
 import React from "react";
 import "../../style/searchbar.css";
 import ForvaltningsElement from "../ForvaltningsKartlag/ForvaltningsElement";
+import backend from "../../Funksjoner/backend";
 
-const SearchBar = props => {
-  let treffliste = props.treffliste;
-  let treffliste_lokalt = props.treffliste_lokalt;
+class SearchBar extends React.Component {
+  state = {
+    treffliste: null,
+    treffliste_lokalt: null,
+    fylker: null,
+    kommuner: null
+  };
 
-  return (
-    <div className="searchbar_container">
-      <div className="searchbar">
-        <input
-          className="searchbarfield"
-          id="searchfield"
-          type="text"
-          placeholder="søk i sideinnhold eller geografisk område"
-          onChange={e => {
-            if (e.target.value) {
-              props.handleSearchBar(e.target.value.toLowerCase());
-            } else {
-              props.handleSearchBar(null);
+  handleRemoveTreffliste = () => {
+    this.setState({
+      treffliste: null,
+      treffliste_lokalt: null
+    });
+  };
+
+  async fetchGeoData() {
+    let kommuner = this.state.kommuner;
+    if (kommuner === null) {
+      await backend.hentKommuner().then(henta_kommuner => {
+        this.setState({
+          kommuner: henta_kommuner
+        });
+      });
+    }
+  }
+
+  handleGeoSelection = geostring => {
+    if (geostring[1] === "Kommune") {
+      backend.hentKommunePolygon(geostring[2]).then(resultat => {
+        let polygon = resultat.omrade.coordinates[0];
+        let minx = 100;
+        let maxy = 0;
+        let maxx = 0;
+        let miny = 100;
+        for (let i in polygon) {
+          let this_item = polygon[i];
+          for (let i in this_item) {
+            let item = this_item[i];
+            if (item[0] < minx) {
+              minx = item[0];
+            } else if (item[0] > maxx) {
+              maxx = item[0];
             }
-          }}
-        />
+            if (item[1] > maxy) {
+              maxy = item[1];
+            } else if (item[1] < miny) {
+              miny = item[1];
+            }
+          }
+        }
+        let mincoord = [minx, miny];
+        let maxcoord = [maxx, maxy];
+        let centercoord = [(minx + maxx) / 2, (miny + maxy) / 2];
+        this.props.handleSetZoomCoordinates(mincoord, maxcoord, centercoord);
+      });
+    }
+  };
 
-        <button
-          onClick={() => {
-            props.handleRemoveTreffliste();
-            props.handleSearchBar(null);
-            document.getElementById("searchfield").value = "";
-          }}
-        >
-          x
-        </button>
+  handleSearchBar = searchTerm => {
+    let counter = 0;
+    let treffliste_lokalt = [];
+    let lag = this.props.kartlag;
+
+    function searchForKey(criteria, counter, lag, searchTerm) {
+      let treffliste_lokalt = [];
+      for (let i in lag) {
+        if (counter >= 5) {
+          break;
+        } else {
+          if (lag[i][criteria]) {
+            let lagstring = lag[i][criteria].toLowerCase();
+            if (lagstring.indexOf(searchTerm) !== -1) {
+              let element = lag[i];
+              element.id = Object.keys(lag)[i];
+              treffliste_lokalt.push(element);
+              counter += 1;
+            }
+          }
+        }
+      }
+      return [treffliste_lokalt, counter];
+    }
+
+    if (searchTerm && searchTerm.length > 0) {
+      let title_search = searchForKey("tittel", counter, lag, searchTerm);
+      treffliste_lokalt = title_search[0];
+      counter = title_search[1];
+      let owner_search = searchForKey("dataeier", counter, lag, searchTerm);
+      treffliste_lokalt = treffliste_lokalt.concat(owner_search[0]);
+      counter += owner_search[1];
+      let theme_search = searchForKey("tema", counter, lag, searchTerm);
+      treffliste_lokalt = treffliste_lokalt.concat(theme_search[0]);
+      counter += theme_search[1];
+    }
+    this.setState({
+      treffliste_lokalt: treffliste_lokalt
+    });
+
+    this.fetchGeoData().then(() => {
+      let counter2 = 0;
+      let kommuner = this.state.kommuner;
+      let treffliste = [];
+      for (let i in kommuner) {
+        if (counter2 >= 5) {
+          break;
+        } else {
+          let treff = kommuner[i].kommunenavn.toLowerCase();
+          if (treff.indexOf(searchTerm) !== -1) {
+            treffliste.push([
+              kommuner[i].kommunenavn,
+              "Kommune",
+              kommuner[i].kommunenummer
+            ]);
+            counter2 += 1;
+          }
+        }
+      }
+      this.setState({
+        treffliste: treffliste
+      });
+    });
+  };
+
+  render() {
+    let treffliste = this.state.treffliste;
+    let treffliste_lokalt = this.state.treffliste_lokalt;
+
+    return (
+      <div className="searchbar_container">
+        <div className="searchbar">
+          <input
+            className="searchbarfield"
+            id="searchfield"
+            type="text"
+            placeholder="søk i sideinnhold eller geografisk område"
+            onChange={e => {
+              if (e.target.value) {
+                this.handleSearchBar(e.target.value.toLowerCase());
+              } else {
+                this.handleSearchBar(null);
+              }
+            }}
+          />
+
+          <button
+            onClick={() => {
+              this.handleRemoveTreffliste();
+              this.handleSearchBar(null);
+              document.getElementById("searchfield").value = "";
+            }}
+          >
+            x
+          </button>
+        </div>
+
+        <div className="treffliste">
+          {treffliste &&
+            treffliste.length > 0 &&
+            treffliste.map(item => {
+              let itemname = item[0] || "";
+              let itemtype = item[1] || "";
+              let itemnr = item[2] || "";
+              return (
+                <button
+                  className="searchbar_item"
+                  key={item}
+                  onClick={() => {
+                    this.handleGeoSelection(item);
+                    this.handleRemoveTreffliste();
+                    document.getElementById("searchfield").value = "";
+                  }}
+                >
+                  <span className="itemname">{itemname} </span>
+                  <span className="itemtype">{itemtype} </span>
+                  <span className="itemnr">{itemnr} </span>
+                </button>
+              );
+            })}
+
+          {treffliste_lokalt &&
+            treffliste_lokalt.length > 0 &&
+            treffliste_lokalt.map(item => {
+              return (
+                <div className="searchbar_item searchbar_local_item_container">
+                  <ForvaltningsElement
+                    kartlag_key={item.id}
+                    kartlag={item}
+                    key={item.id}
+                    onUpdateLayerProp={this.props.onUpdateLayerProp}
+                  />
+                </div>
+              );
+            })}
+        </div>
       </div>
-
-      <div className="treffliste">
-        {treffliste &&
-          props.treffliste.length > 0 &&
-          treffliste.map(item => {
-            let itemname = item[0] || "";
-            let itemtype = item[1] || "";
-            let itemnr = item[2] || "";
-            return (
-              <button
-                className="searchbar_item"
-                key={item}
-                onClick={() => {
-                  props.handleGeoSelection(item);
-                  props.handleRemoveTreffliste();
-                  document.getElementById("searchfield").value = "";
-                }}
-              >
-                <span className="itemname">{itemname} </span>
-                <span className="itemtype">{itemtype} </span>
-                <span className="itemnr">{itemnr} </span>
-              </button>
-            );
-          })}
-
-        {treffliste_lokalt &&
-          props.treffliste_lokalt.length > 0 &&
-          treffliste_lokalt.map(item => {
-            return (
-              <div className="searchbar_item searchbar_local_item_container">
-                <ForvaltningsElement
-                  kartlag_key={item.id}
-                  kartlag={item}
-                  key={item.id}
-                  onUpdateLayerProp={props.onUpdateLayerProp}
-                />
-              </div>
-            );
-          })}
-      </div>
-    </div>
-  );
-};
+    );
+  }
+}
 
 export default SearchBar;
